[BREAKING CHANGES] add basic input dispatcher support.
It will be used by resident-midi-keyboard as it is going to implement
virtual MIDI "input keyboard" support (output in Android MIDI API wording).

To make it happen, we will need some functionality that can dispatch
MIDI output events from this DiatonicLiveMidiKeyboard to the registered
listeners. MidiOutput implementation can just register them and dispatch
whatever it received from the listener to its own `outputPorReceivers`.

This technically breaks the MidiDeviceAccessScope API that hides
ktmidi MidiOutput (currentOutput). It should be almost harmless but we'd
raise the version to 0.3 as it is also a significant addition.