workaround for #3 - forcibly delay midi2 protocol switching at UI by 500msec.
The "correct" implementation would have a bi-directional connections with
the MIDI OUT device, send UMP Stream Configuration Request message and
wait for Stream Configuration Notification response. Since we do not really
expect that, things are hacky anyways.

(We could still implement a valid asynchronous UI event notification and
disable any MIDI events while while we are waiting for the "response", but
people should not expect that precise behavior on this app at least at this
state.)